How descriptions are obtained by the MDE:
Following the route of the algorithm:

First we must solve our equation with the solver.
solver.add(equation);
solver.solve();

If it can describe the equation, get get the describer to give it to us

String description = describer.getDescriptions("mode of description goes 

here");

getDescriptions("mode")  will set the mode to be mode and then call 

getDescriptions() and return the string from that.


getDescriptions() is where this becomes tricky.  It cosists of grabbing xml 

from the xsl files and then transforming it into a String.  The 

transformation does not bear mentioning at the moment;  the grabbing is 

what is pertains to us.

The pertinent call in all of this is 

xmlToTransform = getFeatureXML(); // magic

However, to fully understand what is going on, we must remove the magic: 

http://abstrusegoose.com/98


private String getFeatureXML() does the following.

//Build the XML stream:

XML stream contains the following: <MDE> [features] </MDE>

// Get the XML block from the features.
        Solution solution;
        SolvedGraph features;

we then iterate thru and append features to the stream
features = solution.getFeatures();

appending is not a mystery, so we now we need to figure out what happens in 

solution.getFeatures();


solution.getFeatures() ->
 
Get the SolvedGraph associated with this Solution. The graph's describable 

features can be accessed from the SolvedGraph object.  

SolvedGraphs can grab the features from its feature tree and associated