MDE: Started creating interfaces for the various properties a graph may have.  The first one is SlopeGraph.
The MDE provides very little access to the properties hashtable.  The properties can be easily added but are difficult to extract.  There is no "get" method!  This is now changed: MdeFeatureNodes and SolvedGraphs now provide ways to access their features.  A path and key are required.  If the key is not found, a NullPointerException is thrown.