Fixed a bug where we were crashing on large coefficients.
The problem was in NumberModel.java.  The code was formating numbers that were in the thousands+ with commas, then trying to create a Double object with that, ie:

approximateDecimalValue = n.format(decimalValue);
double d = new Double(approximateDecimalValue which is something like 24,324,234.35675672).doubleValue();

I fixed that by removing the commas.

I am marking this as closed since we fixed the "crash on large problems bug."